carfront  Snack

+++ carfront/src/component/Carlist.js
@@ -1,8 +1,10 @@
 import React, {useEffect, useState} from 'react';
 import {SERVER_URL} from "../constants";
 import {DataGrid} from '@mui/x-data-grid';
+import Snackbar from '@mui/material/Snackbar';
 
 function CarList() {
+    const [open, setOpen] = useState(false);
     const [cars, setCars] = useState([]);
     const columns = [
         {field: 'brand', headerName: 'Brand', width: 200},
@@ -26,9 +28,18 @@ function CarList() {
             .catch(err => console.error(err));
 
     const onDelClick = (url) => {
-        fetch(url, {method: 'DELETE'})
-            .then(response => fetchCars())
-            .catch(err => console.error(err))
+        if (window.confirm("Are you sure to delete?")) {
+            fetch(url, {method: 'DELETE'})
+                .then(response => {
+                    if (response.ok) {
+                        fetchCars();
+                        setOpen(true);
+                    } else {
+                        alert('Something went wrong!');
+                    }
+                })
+                .catch(err => console.error(err))
+        }
     }
 
     useEffect(() => {
@@ -39,8 +50,15 @@ function CarList() {
             <DataGrid
                 rows={cars}
                 columns={columns}
+                disableSelectionOnClick={true}
                 getRowId={row => row._links.self.href}/>
+            <Snackbar
+                open={open}
+                autoHideDuration={2000}
+                onClose={() => setOpen(false)}
+                message="Car deleted"/>
         </div>
+
     );
 }
